only include public interfaces in opc.h
