+++ driver/opc/util.h
@@ -25,6 +25,116 @@
 #include <iostream>
 
 namespace opc {
+/// @brief the configuration for an OPC UA connection.
+struct ConnectionConfig {
+    /// @brief the endpoint of the OPC UA server.
+    std::string endpoint;
+    /// @brief the username to use for authentication. Not required.
+    std::string username;
+    /// @brief the password to use for authentication. Not required.
+    std::string password;
+    /// @brief the security mode.
+    std::string security_mode;
+    /// @brief the security policy.
+    std::string security_policy;
+    /// @brief the client certificate used to sign and encrypt messages. Only required
+    /// if the security policy is not "None".
+    std::string client_cert;
+    /// @brief the client private key used to sign and encrypt messages. Only required
+    /// if the security policy is not "None".
+    std::string client_private_key;
+    /// @brief a trusted server certificate. Only req
+    std::string server_cert;
+
+    ConnectionConfig() = default;
+
+    explicit ConnectionConfig(
+        config::Parser parser
+    ) : endpoint(parser.required<std::string>("endpoint")),
+        username(parser.optional<std::string>("username", "")),
+        password(parser.optional<std::string>("password", "")),
+        security_mode(parser.optional<std::string>("security_mode", "None")),
+        security_policy(parser.optional<std::string>("security_policy", "None")),
+        client_cert(parser.optional<std::string>("client_certificate", "")),
+        client_private_key(parser.optional<std::string>("client_private_key", "")),
+        server_cert(parser.optional<std::string>("server_certificate", "")) {
+    }
+
+    json toJSON() const {
+        return {
+            {"endpoint",           endpoint},
+            {"username",           username},
+            {"password",           password},
+            {"security_mode",      security_mode},
+            {"security_policy",    security_policy},
+            {"client_certificate", client_cert},
+            {"client_private_key", client_private_key}
+        };
+    }
+};
+
+struct DeviceNodeProperties {
+    synnax::DataType data_type;
+    std::string node_class;
+    std::string name;
+    std::string node_id;
+    bool is_array;
+
+    DeviceNodeProperties(
+        synnax::DataType data_type,
+        std::string name,
+        std::string node_id,
+        std::string node_class,
+        bool is_array
+    ) : data_type(data_type), name(name), node_id(node_id), node_class(node_class), is_array(is_array) {
+    }
+
+    explicit DeviceNodeProperties(config::Parser parser) : data_type(
+        synnax::DataType(parser.required<std::string>("data_type"))),
+                                                           name(parser.required<std::string>("name")),
+                                                           node_id(parser.required<std::string>("node_id")),
+                                                           is_array(parser.optional<bool>("is_array", false)) {
+    }
+
+    json toJSON() const {
+        return {
+            {"data_type",  data_type.name()},
+            {"name",       name},
+            {"node_id",    node_id},
+            {"node_class", node_class},
+            {"is_array",   is_array}
+        };
+    }
+};
+
+struct DeviceProperties {
+    ConnectionConfig connection;
+    std::vector<DeviceNodeProperties> channels;
+
+    DeviceProperties(
+        ConnectionConfig connection,
+        std::vector<DeviceNodeProperties> channels
+    ) : connection(connection), channels(channels) {
+    }
+
+    explicit DeviceProperties(
+        config::Parser parser
+    ) : connection(parser.child("connection")),
+        channels({}) {
+        parser.iter("channels", [&](const config::Parser &cb) {
+            channels.emplace_back(cb);
+        });
+    }
+
+    json toJSON() const {
+        json j;
+        j["connection"] = connection.toJSON();
+        j["channels"] = json::array();
+        for (const auto &ch: channels)
+            j["channels"].push_back(ch.toJSON());
+        return j;
+    }
+};
     using ClientDeleter = void (*)(UA_Client *);
